Code review updates
git-svn-id: https://svn.code.sf.net/p/cppcms/code/boost_locale/trunk@1794 831565cc-8e21-4ad5-8334-98019c97bd36

+++ boost/locale/date_time.hpp
@@ -417,7 +417,7 @@ namespace boost {
             ///
             /// Get item at position \a n the set, n should be in range [0,size)
             ///
-            date_time_period const &operator[](unsigned n) const 
+            date_time_period const &operator[](size_t n) const 
             {
                 if(n >= size())
                     throw std::out_of_range("Invalid index to date_time_period");
